+++ VTK/Graphics/vtkPointSource.cxx
@@ -44,8 +44,8 @@ vtkPointSource::vtkPointSource(vtkIdType numPts)
 void vtkPointSource::Execute()
 {
   vtkIdType i;
-  float theta, rho, cosphi, sinphi, radius;
-  float x[3];
+  double theta, rho, cosphi, sinphi, radius;
+  double x[3];
   vtkPoints *newPoints;
   vtkCellArray *newVerts;
   vtkPolyData *output = this->GetOutput();
